+++ source/buttons/tinker.js
@@ -4,9 +4,9 @@ const { getGearProperty } = require('../gear/_gearDictionary');
 const { getAdventure } = require('../orcustrators/adventureOrcustrator');
 const { SAFE_DELIMITER } = require('../constants');
 
-const mainId = "upgrade";
+const mainId = "tinker";
 module.exports = new ButtonWrapper(mainId, 3000,
-	/** Present the user with an opportunity to upgrade a piece of gear */
+	/** Present the user with an opportunity to tinker with a piece of gear */
 	(interaction, args) => {
 		const adventure = getAdventure(interaction.channelId);
 		const delver = adventure?.delvers.find(delver => delver.id === interaction.user.id);
@@ -22,26 +22,26 @@ module.exports = new ButtonWrapper(mainId, 3000,
 
 		const options = [];
 		delver.gear.forEach((gear, index) => {
-			const upgrades = getGearProperty(gear.name, "upgrades");
-			if (upgrades.length > 0) {
+			const sidegrades = getGearProperty(gear.name, "sidegrades");
+			if (sidegrades.length > 0) {
 				options.push({
 					label: gear.name,
-					description: `Results: ${upgrades.join(", ")}`,
+					description: `Results: ${sidegrades.join(", ")}`,
 					value: `${gear.name}${SAFE_DELIMITER}${index}`
 				})
 			}
 		})
 		if (options.length < 1) {
-			interaction.reply({ content: "You don't have any gear that can be upgraded.", ephemeral: true });
+			interaction.reply({ content: "You don't have any gear that can be tinkered with.", ephemeral: true });
 			return;
 		}
 
 		interaction.reply({
-			content: "You can use 1 room action to upgrade a piece of gear.",
+			content: "You can use 1 room action to change a piece of gear to a different upgrade.",
 			components: [
 				new ActionRowBuilder().addComponents(
-					new StringSelectMenuBuilder().setCustomId("randomupgrade")
-						.setPlaceholder("Pick a piece of gear to randomly upgrade...")
+					new StringSelectMenuBuilder().setCustomId("randomtinker")
+						.setPlaceholder("Pick a piece of gear to randomly tinker with...")
 						.setOptions(options)
 				)
 			],