+++ source/selects/buygear.js
@@ -1,8 +1,8 @@
 const { ActionRowBuilder, ButtonBuilder, ButtonStyle } = require('discord.js');
-const { SelectWrapper, Gear } = require('../classes');
+const { SelectWrapper } = require('../classes');
 const { SAFE_DELIMITER } = require('../constants');
 const { getAdventure, setAdventure } = require('../orcustrators/adventureOrcustrator');
-const { getGearProperty } = require('../gear/_gearDictionary');
+const { buildGearRecord } = require('../gear/_gearDictionary');
 const { renderRoom } = require('../util/embedUtil');
 
 const mainId = "buygear";
@@ -23,7 +23,7 @@ module.exports = new SelectWrapper(mainId, 3000,
 				if (delver.gear.length < adventure.getGearCapacity()) {
 					adventure.gold -= cost;
 					adventure.room.resources[name].count--;
-					delver.gear.push(new Gear(name, getGearProperty(name, "maxDurability"), getGearProperty(name, "maxHP"), getGearProperty(name, "speed"), getGearProperty(name, "critRate"), getGearProperty(name, "poise")));
+					delver.gear.push(buildGearRecord(name, "max"));
 					interaction.message.edit(renderRoom(adventure, interaction.channel));
 					interaction.reply({ content: `${interaction.member.displayName} buys a ${name} for ${cost}g.` });
 					setAdventure(adventure);