+++ source/selects/treasure.js
@@ -1,7 +1,7 @@
 const { ActionRowBuilder, ButtonBuilder, ButtonStyle } = require('discord.js');
-const { SelectWrapper, Gear } = require('../classes');
+const { SelectWrapper } = require('../classes');
 const { getAdventure, setAdventure } = require('../orcustrators/adventureOrcustrator');
-const { getGearProperty } = require('../gear/_gearDictionary');
+const { buildGearRecord } = require('../gear/_gearDictionary');
 const { SAFE_DELIMITER, ZERO_WIDTH_WHITESPACE } = require('../constants');
 const { renderRoom } = require('../util/embedUtil');
 
@@ -47,7 +47,7 @@ module.exports = new SelectWrapper(mainId, 2000,
 					break;
 				case "gear":
 					if (delver.gear.length < adventure.getGearCapacity()) {
-						delver.gear.push(new Gear(name, getGearProperty(name, "maxDurability"), getGearProperty(name, "maxHP"), getGearProperty(name, "speed"), getGearProperty(name, "critRate"), getGearProperty(name, "poise")));
+						delver.gear.push(buildGearRecord(name, "max"));
 						adventure.room.resources[name].count = Math.max(count - 1, 0);
 						result = {
 							content: `${interaction.member.displayName} takes a ${name}. There are ${count - 1} remaining.`